--- src/app/admin.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AdminService } from './admin.service';
-
-describe('AdminService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: AdminService = TestBed.get(AdminService);
-    expect(service).toBeTruthy();
-  });
-});
