+++ src/app/admin/get-events/get-events.component.ts
@@ -1,5 +1,5 @@
 import { Router } from '@angular/router';
-import { AdminService } from './../../admin.service';
+import { AdminService } from '../../shared/services/admin.service';
 import { Component } from '@angular/core';
 
 @Component({
