+++ src/app/shared/services/admin.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { environment } from './../environments/environment';
+import { environment } from './../../../environments/environment';
 import { HttpClient } from '@angular/common/http';
 
 @Injectable({
